requirements.txt
bcrypt
celery
coverage
cryptography==1.2.3
flask
flask-breadcrumbs
flask-cache
flask-celery-helper
flask-csp
flask-login
flask-migrate
flask-sqlalchemy
flask-script
flask-wtf
gitdb
gitpython
gunicorn
markdown2
mysql-python
onetimepass
passlib
paramiko
pathlib
pillow
pycrypto
pyqrcode
pytest
pyyaml
rauth
redis
requests
sqlalchemy==1.1.0b3
wtforms_components